Fix AI ability to attack from the moat, move auxiliary pathfinding-related functions from AI to Board
