+++ src/fheroes2/battle/battle_board.h
@@ -95,9 +95,16 @@ namespace Battle
         static Indexes GetAroundIndexes( const Unit & );
         static Indexes GetMoveWideIndexes( s32, bool reflect );
         static bool isValidMirrorImageIndex( s32, const Unit * );
+
+        // Checks that the attacker is able (in principle) to attack from the given cell
         static bool CanAttackUnitFromCell( const Unit & attacker, const int32_t from );
+        // Checks that the attacker is able to attack the target from the given cell
+        static bool CanAttackUnitFromCell( const Unit & attacker, const Unit & target, const int32_t from );
 
         static Indexes GetAdjacentEnemies( const Unit & unit );
+
+        // Finds the cell nearest to the target and reachable for the unit
+        static int32_t FindNearestReachableCell( const int32_t target, const Unit & unit );
 
         // Handles the situation when the destination cell is on the border of the cell
         // space reachable for the unit and it should be the tail cell of this unit
